refactor(core): extract static concerns from CoreToolScheduler
This change extracts static types and utility functions from CoreToolScheduler to improve maintainability and testability.

- Extracted types to 'packages/core/src/scheduler/types.ts'
- Moved 'convertToFunctionResponse' to 'packages/core/src/utils/generateContentResponseUtilities.ts'
- Moved 'truncateAndSaveToFile' (renamed to 'saveTruncatedContent') to 'packages/core/src/utils/fileUtils.ts'
- Moved 'getToolSuggestion' to 'packages/core/src/utils/tool-utils.ts'
- Updated tests and imports accordingly.
