+++ packages/core/src/core/coreToolScheduler.ts
@@ -27,21 +27,19 @@ import {
   ToolOutputTruncatedEvent,
   runInDevTraceSpan,
 } from '../index.js';
-import { READ_FILE_TOOL_NAME, SHELL_TOOL_NAME } from '../tools/tool-names.js';
-import type { Part, PartListUnion } from '@google/genai';
-import { supportsMultimodalFunctionResponse } from '../config/models.js';
+import { SHELL_TOOL_NAME } from '../tools/tool-names.js';
 import type { ModifyContext } from '../tools/modifiable-tool.js';
 import {
   isModifiableDeclarativeTool,
   modifyWithEditor,
 } from '../tools/modifiable-tool.js';
 import * as Diff from 'diff';
-import * as fs from 'node:fs/promises';
-import * as path from 'node:path';
 import { SHELL_TOOL_NAMES } from '../utils/shell-utils.js';
-import { doesToolInvocationMatch } from '../utils/tool-utils.js';
+import {
+  doesToolInvocationMatch,
+  getToolSuggestion,
+} from '../utils/tool-utils.js';
 import { isShellInvocationAllowlisted } from '../utils/shell-permissions.js';
-import levenshtein from 'fast-levenshtein';
 import { ShellToolInvocation } from '../tools/shell.js';
 import type { ToolConfirmationRequest } from '../confirmation-bus/types.js';
 import { MessageBusType } from '../confirmation-bus/types.js';
@@ -50,220 +48,41 @@ import {
   fireToolNotificationHook,
   executeToolWithHooks,
 } from './coreToolHookTriggers.js';
-import { debugLogger } from '../utils/debugLogger.js';
+import {
+  type ToolCall,
+  type ValidatingToolCall,
+  type ScheduledToolCall,
+  type ErroredToolCall,
+  type SuccessfulToolCall,
+  type ExecutingToolCall,
+  type CancelledToolCall,
+  type WaitingToolCall,
+  type Status,
+  type CompletedToolCall,
+  type ConfirmHandler,
+  type OutputUpdateHandler,
+  type AllToolCallsCompleteHandler,
+  type ToolCallsUpdateHandler,
+} from '../scheduler/types.js';
+import { saveTruncatedContent } from '../utils/fileUtils.js';
+import { convertToFunctionResponse } from '../utils/generateContentResponseUtilities.js';
 
-export type ValidatingToolCall = {
-  status: 'validating';
-  request: ToolCallRequestInfo;
-  tool: AnyDeclarativeTool;
-  invocation: AnyToolInvocation;
-  startTime?: number;
-  outcome?: ToolConfirmationOutcome;
+export type {
+  ToolCall,
+  ValidatingToolCall,
+  ScheduledToolCall,
+  ErroredToolCall,
+  SuccessfulToolCall,
+  ExecutingToolCall,
+  CancelledToolCall,
+  WaitingToolCall,
+  Status,
+  CompletedToolCall,
+  ConfirmHandler,
+  OutputUpdateHandler,
+  AllToolCallsCompleteHandler,
+  ToolCallsUpdateHandler,
 };
-
-export type ScheduledToolCall = {
-  status: 'scheduled';
-  request: ToolCallRequestInfo;
-  tool: AnyDeclarativeTool;
-  invocation: AnyToolInvocation;
-  startTime?: number;
-  outcome?: ToolConfirmationOutcome;
-};
-
-export type ErroredToolCall = {
-  status: 'error';
-  request: ToolCallRequestInfo;
-  response: ToolCallResponseInfo;
-  tool?: AnyDeclarativeTool;
-  durationMs?: number;
-  outcome?: ToolConfirmationOutcome;
-};
-
-export type SuccessfulToolCall = {
-  status: 'success';
-  request: ToolCallRequestInfo;
-  tool: AnyDeclarativeTool;
-  response: ToolCallResponseInfo;
-  invocation: AnyToolInvocation;
-  durationMs?: number;
-  outcome?: ToolConfirmationOutcome;
-};
-
-export type ExecutingToolCall = {
-  status: 'executing';
-  request: ToolCallRequestInfo;
-  tool: AnyDeclarativeTool;
-  invocation: AnyToolInvocation;
-  liveOutput?: string | AnsiOutput;
-  startTime?: number;
-  outcome?: ToolConfirmationOutcome;
-  pid?: number;
-};
-
-export type CancelledToolCall = {
-  status: 'cancelled';
-  request: ToolCallRequestInfo;
-  response: ToolCallResponseInfo;
-  tool: AnyDeclarativeTool;
-  invocation: AnyToolInvocation;
-  durationMs?: number;
-  outcome?: ToolConfirmationOutcome;
-};
-
-export type WaitingToolCall = {
-  status: 'awaiting_approval';
-  request: ToolCallRequestInfo;
-  tool: AnyDeclarativeTool;
-  invocation: AnyToolInvocation;
-  confirmationDetails: ToolCallConfirmationDetails;
-  startTime?: number;
-  outcome?: ToolConfirmationOutcome;
-};
-
-export type Status = ToolCall['status'];
-
-export type ToolCall =
-  | ValidatingToolCall
-  | ScheduledToolCall
-  | ErroredToolCall
-  | SuccessfulToolCall
-  | ExecutingToolCall
-  | CancelledToolCall
-  | WaitingToolCall;
-
-export type CompletedToolCall =
-  | SuccessfulToolCall
-  | CancelledToolCall
-  | ErroredToolCall;
-
-export type ConfirmHandler = (
-  toolCall: WaitingToolCall,
-) => Promise<ToolConfirmationOutcome>;
-
-export type OutputUpdateHandler = (
-  toolCallId: string,
-  outputChunk: string | AnsiOutput,
-) => void;
-
-export type AllToolCallsCompleteHandler = (
-  completedToolCalls: CompletedToolCall[],
-) => Promise<void>;
-
-export type ToolCallsUpdateHandler = (toolCalls: ToolCall[]) => void;
-
-/**
- * Formats tool output for a Gemini FunctionResponse.
- */
-function createFunctionResponsePart(
-  callId: string,
-  toolName: string,
-  output: string,
-): Part {
-  return {
-    functionResponse: {
-      id: callId,
-      name: toolName,
-      response: { output },
-    },
-  };
-}
-
-export function convertToFunctionResponse(
-  toolName: string,
-  callId: string,
-  llmContent: PartListUnion,
-  model: string,
-): Part[] {
-  if (typeof llmContent === 'string') {
-    return [createFunctionResponsePart(callId, toolName, llmContent)];
-  }
-
-  const parts = toParts(llmContent);
-
-  // Separate text from binary types
-  const textParts: string[] = [];
-  const inlineDataParts: Part[] = [];
-  const fileDataParts: Part[] = [];
-
-  for (const part of parts) {
-    if (part.text !== undefined) {
-      textParts.push(part.text);
-    } else if (part.inlineData) {
-      inlineDataParts.push(part);
-    } else if (part.fileData) {
-      fileDataParts.push(part);
-    } else if (part.functionResponse) {
-      if (parts.length > 1) {
-        debugLogger.warn(
-          'convertToFunctionResponse received multiple parts with a functionResponse. Only the functionResponse will be used, other parts will be ignored',
-        );
-      }
-      // Handle passthrough case
-      return [
-        {
-          functionResponse: {
-            id: callId,
-            name: toolName,
-            response: part.functionResponse.response,
-          },
-        },
-      ];
-    }
-    // Ignore other part types
-  }
-
-  // Build the primary response part
-  const part: Part = {
-    functionResponse: {
-      id: callId,
-      name: toolName,
-      response: textParts.length > 0 ? { output: textParts.join('\n') } : {},
-    },
-  };
-
-  const isMultimodalFRSupported = supportsMultimodalFunctionResponse(model);
-  const siblingParts: Part[] = [...fileDataParts];
-
-  if (inlineDataParts.length > 0) {
-    if (isMultimodalFRSupported) {
-      // Nest inlineData if supported by the model
-      (part.functionResponse as unknown as { parts: Part[] }).parts =
-        inlineDataParts;
-    } else {
-      // Otherwise treat as siblings
-      siblingParts.push(...inlineDataParts);
-    }
-  }
-
-  // Add descriptive text if the response object is empty but we have binary content
-  if (
-    textParts.length === 0 &&
-    (inlineDataParts.length > 0 || fileDataParts.length > 0)
-  ) {
-    const totalBinaryItems = inlineDataParts.length + fileDataParts.length;
-    part.functionResponse!.response = {
-      output: `Binary content provided (${totalBinaryItems} item(s)).`,
-    };
-  }
-
-  if (siblingParts.length > 0) {
-    return [part, ...siblingParts];
-  }
-
-  return [part];
-}
-
-function toParts(input: PartListUnion): Part[] {
-  const parts: Part[] = [];
-  for (const part of Array.isArray(input) ? input : [input]) {
-    if (typeof part === 'string') {
-      parts.push({ text: part });
-    } else if (part) {
-      parts.push(part);
-    }
-  }
-  return parts;
-}
 
 const createErrorResponse = (
   request: ToolCallRequestInfo,
@@ -285,70 +104,6 @@ const createErrorResponse = (
   errorType,
   contentLength: error.message.length,
 });
-
-export async function truncateAndSaveToFile(
-  content: string,
-  callId: string,
-  projectTempDir: string,
-  threshold: number,
-  truncateLines: number,
-): Promise<{ content: string; outputFile?: string }> {
-  if (content.length <= threshold) {
-    return { content };
-  }
-
-  let lines = content.split('\n');
-  let fileContent = content;
-
-  // If the content is long but has few lines, wrap it to enable line-based truncation.
-  if (lines.length <= truncateLines) {
-    const wrapWidth = 120; // A reasonable width for wrapping.
-    const wrappedLines: string[] = [];
-    for (const line of lines) {
-      if (line.length > wrapWidth) {
-        for (let i = 0; i < line.length; i += wrapWidth) {
-          wrappedLines.push(line.substring(i, i + wrapWidth));
-        }
-      } else {
-        wrappedLines.push(line);
-      }
-    }
-    lines = wrappedLines;
-    fileContent = lines.join('\n');
-  }
-
-  const head = Math.floor(truncateLines / 5);
-  const beginning = lines.slice(0, head);
-  const end = lines.slice(-(truncateLines - head));
-  const truncatedContent =
-    beginning.join('\n') + '\n... [CONTENT TRUNCATED] ...\n' + end.join('\n');
-
-  // Sanitize callId to prevent path traversal.
-  const safeFileName = `${path.basename(callId)}.output`;
-  const outputFile = path.join(projectTempDir, safeFileName);
-  try {
-    await fs.writeFile(outputFile, fileContent);
-
-    return {
-      content: `Tool output was too large and has been truncated.
-The full output has been saved to: ${outputFile}
-To read the complete output, use the ${READ_FILE_TOOL_NAME} tool with the absolute file path above. For large files, you can use the offset and limit parameters to read specific sections:
-- ${READ_FILE_TOOL_NAME} tool with offset=0, limit=100 to see the first 100 lines
-- ${READ_FILE_TOOL_NAME} tool with offset=N to skip N lines from the beginning
-- ${READ_FILE_TOOL_NAME} tool with limit=M to read only M lines at a time
-The truncated output below shows the beginning and end of the content. The marker '... [CONTENT TRUNCATED] ...' indicates where content was removed.
-This allows you to efficiently examine different parts of the output without loading the entire file.
-Truncated part of the output:
-${truncatedContent}`,
-      outputFile,
-    };
-  } catch (_error) {
-    return {
-      content:
-        truncatedContent + `\n[Note: Could not save full output to file]`,
-    };
-  }
-}
 
 interface CoreToolSchedulerOptions {
   config: Config;
@@ -656,40 +411,6 @@ export class CoreToolScheduler {
     }
   }
 
-  /**
-   * Generates a suggestion string for a tool name that was not found in the registry.
-   * It finds the closest matches based on Levenshtein distance.
-   * @param unknownToolName The tool name that was not found.
-   * @param topN The number of suggestions to return. Defaults to 3.
-   * @returns A suggestion string like " Did you mean 'tool'?" or " Did you mean one of: 'tool1', 'tool2'?", or an empty string if no suggestions are found.
-   */
-  private getToolSuggestion(unknownToolName: string, topN = 3): string {
-    const allToolNames = this.config.getToolRegistry().getAllToolNames();
-
-    const matches = allToolNames.map((toolName) => ({
-      name: toolName,
-      distance: levenshtein.get(unknownToolName, toolName),
-    }));
-
-    matches.sort((a, b) => a.distance - b.distance);
-
-    const topNResults = matches.slice(0, topN);
-
-    if (topNResults.length === 0) {
-      return '';
-    }
-
-    const suggestedNames = topNResults
-      .map((match) => `"${match.name}"`)
-      .join(', ');
-
-    if (topNResults.length > 1) {
-      return ` Did you mean one of: ${suggestedNames}?`;
-    } else {
-      return ` Did you mean ${suggestedNames}?`;
-    }
-  }
-
   schedule(
     request: ToolCallRequestInfo | ToolCallRequestInfo[],
     signal: AbortSignal,
@@ -784,7 +505,10 @@ export class CoreToolScheduler {
             .getToolRegistry()
             .getTool(reqInfo.name);
           if (!toolInstance) {
-            const suggestion = this.getToolSuggestion(reqInfo.name);
+            const suggestion = getToolSuggestion(
+              reqInfo.name,
+              this.config.getToolRegistry().getAllToolNames(),
+            );
             const errorMessage = `Tool "${reqInfo.name}" not found in registry. Tools must use the exact names that are registered.${suggestion}`;
             return {
               status: 'error',
@@ -1222,7 +946,7 @@ export class CoreToolScheduler {
                   const threshold =
                     this.config.getTruncateToolOutputThreshold();
                   const lines = this.config.getTruncateToolOutputLines();
-                  const truncatedResult = await truncateAndSaveToFile(
+                  const truncatedResult = await saveTruncatedContent(
                     content,
                     callId,
                     this.config.storage.getProjectTempDir(),
