+++ packages/core/src/scheduler/types.ts
@@ -1,0 +1,114 @@
+/**
+ * @license
+ * Copyright 2025 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import type {
+  ToolCallRequestInfo,
+  ToolCallResponseInfo,
+  ToolCallConfirmationDetails,
+  AnyDeclarativeTool,
+  AnyToolInvocation,
+  AnsiOutput,
+  ToolConfirmationOutcome,
+} from '../index.js';
+
+export type ValidatingToolCall = {
+  status: 'validating';
+  request: ToolCallRequestInfo;
+  tool: AnyDeclarativeTool;
+  invocation: AnyToolInvocation;
+  startTime?: number;
+  outcome?: ToolConfirmationOutcome;
+};
+
+export type ScheduledToolCall = {
+  status: 'scheduled';
+  request: ToolCallRequestInfo;
+  tool: AnyDeclarativeTool;
+  invocation: AnyToolInvocation;
+  startTime?: number;
+  outcome?: ToolConfirmationOutcome;
+};
+
+export type ErroredToolCall = {
+  status: 'error';
+  request: ToolCallRequestInfo;
+  response: ToolCallResponseInfo;
+  tool?: AnyDeclarativeTool;
+  durationMs?: number;
+  outcome?: ToolConfirmationOutcome;
+};
+
+export type SuccessfulToolCall = {
+  status: 'success';
+  request: ToolCallRequestInfo;
+  tool: AnyDeclarativeTool;
+  response: ToolCallResponseInfo;
+  invocation: AnyToolInvocation;
+  durationMs?: number;
+  outcome?: ToolConfirmationOutcome;
+};
+
+export type ExecutingToolCall = {
+  status: 'executing';
+  request: ToolCallRequestInfo;
+  tool: AnyDeclarativeTool;
+  invocation: AnyToolInvocation;
+  liveOutput?: string | AnsiOutput;
+  startTime?: number;
+  outcome?: ToolConfirmationOutcome;
+  pid?: number;
+};
+
+export type CancelledToolCall = {
+  status: 'cancelled';
+  request: ToolCallRequestInfo;
+  response: ToolCallResponseInfo;
+  tool: AnyDeclarativeTool;
+  invocation: AnyToolInvocation;
+  durationMs?: number;
+  outcome?: ToolConfirmationOutcome;
+};
+
+export type WaitingToolCall = {
+  status: 'awaiting_approval';
+  request: ToolCallRequestInfo;
+  tool: AnyDeclarativeTool;
+  invocation: AnyToolInvocation;
+  confirmationDetails: ToolCallConfirmationDetails;
+  startTime?: number;
+  outcome?: ToolConfirmationOutcome;
+};
+
+export type Status = ToolCall['status'];
+
+export type ToolCall =
+  | ValidatingToolCall
+  | ScheduledToolCall
+  | ErroredToolCall
+  | SuccessfulToolCall
+  | ExecutingToolCall
+  | CancelledToolCall
+  | WaitingToolCall;
+
+export type CompletedToolCall =
+  | SuccessfulToolCall
+  | CancelledToolCall
+  | ErroredToolCall;
+
+export type ConfirmHandler = (
+  toolCall: WaitingToolCall,
+) => Promise<ToolConfirmationOutcome>;
+
+export type OutputUpdateHandler = (
+  toolCallId: string,
+  outputChunk: string | AnsiOutput,
+) => void;
+
+export type AllToolCallsCompleteHandler = (
+  completedToolCalls: CompletedToolCall[],
+) => Promise<void>;
+
+export type ToolCallsUpdateHandler = (toolCalls: ToolCall[]) => void;
